Crate pallet_sub

Sub Module

Config

Call

Create Sub tokens from NFT.

Terminology

Sub Token: Lock NFT to this module then create new collection and tokens.

Recover: Restore Sub Token to NFT.

Interface

Dispatchable Functions

create - Transfer NFT to this module, and then create a new collection.

recover - Transfer the locked NFT to the account that has all the sub tokens, and destroy the sub tokens.

mint_non_fungible - Mint one or a batch of SubNFTs

mint_fungible - Mint some SubFTs

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

Module

SubTokenCreator
The set of SubToken creators. subtoken_collection => creator

SubTokens
Record the collection_id of the SubToken corresponding to the locked NFT subtoken_collection => nft(collection_id, start_idx)

Enums

Call
Dispatchable calls.

Error
Errors inform users that something went wrong.

RawEvent
Events for this module.

Traits

Config

Type Definitions

Event
RawEvent specialized for the configuration Config